Move diagram export to plugins
It's a plugin, not a (core) service.
