+++ pyproject.toml
@@ -60,7 +60,6 @@ gaphorconvert = 'gaphor.tools.gaphorconvert:main'
 "element_factory" = "gaphor.UML.elementfactory:ElementFactory"
 "file_manager" = "gaphor.ui.filemanager:FileManager"
 "recent_files" = "gaphor.ui.recentfiles:RecentFiles"
-"diagram_export_manager" = "gaphor.services.diagramexportmanager:DiagramExportManager"
 "main_window" = "gaphor.ui.mainwindow:MainWindow"
 "preferences" = "gaphor.ui.preferences:Preferences"
 "export_menu" = "gaphor.ui.menufragment:MenuFragment"
@@ -68,16 +67,17 @@ gaphorconvert = 'gaphor.tools.gaphorconvert:main'
 "tools_menu" = "gaphor.ui.menufragment:MenuFragment"
 "copy" = "gaphor.services.copyservice:CopyService"
 "sanitizer" = "gaphor.services.sanitizerservice:SanitizerService"
-"xmi_export" = "gaphor.plugins.xmiexport:XMIExport"
-"diagram_layout" = "gaphor.plugins.diagramlayout:DiagramLayout"
-"pynsource" = "gaphor.plugins.pynsource:PyNSource"
-"alignment" = "gaphor.plugins.alignment:Alignment"
 "help" = "gaphor.services.helpservice:HelpService"
 "namespace" = "gaphor.ui.mainwindow:Namespace"
 "toolbox" = "gaphor.ui.mainwindow:Toolbox"
 "diagrams" = "gaphor.ui.mainwindow:Diagrams"
 "consolewindow" = "gaphor.ui.consolewindow:ConsoleWindow"
 "elementeditor" = "gaphor.ui.elementeditor:ElementEditor"
+"diagram_export" = "gaphor.plugins.diagramexport:DiagramExport"
+"xmi_export" = "gaphor.plugins.xmiexport:XMIExport"
+"diagram_layout" = "gaphor.plugins.diagramlayout:DiagramLayout"
+"pynsource" = "gaphor.plugins.pynsource:PyNSource"
+"alignment" = "gaphor.plugins.alignment:Alignment"
 
 [build-system]
 requires = ["poetry==1.0.0b1"]
