+++ setup.py
@@ -116,17 +116,17 @@ setup(
             "tools_menu = gaphor.ui.menufragment:MenuFragment",
             "copy = gaphor.services.copyservice:CopyService",
             "sanitizer = gaphor.services.sanitizerservice:SanitizerService",
-            "xmi_export = gaphor.plugins.xmiexport:XMIExport",
-            "diagram_layout = gaphor.plugins.diagramlayout:DiagramLayout",
-            "pynsource = gaphor.plugins.pynsource:PyNSource",
-            "alignment = gaphor.plugins.alignment:Alignment",
             "help = gaphor.services.helpservice:HelpService",
             "namespace = gaphor.ui.mainwindow:Namespace",
             "toolbox = gaphor.ui.mainwindow:Toolbox",
             "diagrams = gaphor.ui.mainwindow:Diagrams",
             "consolewindow = gaphor.ui.consolewindow:ConsoleWindow",
             "elementeditor = gaphor.ui.elementeditor:ElementEditor",
-            "diagram_export_manager = gaphor.services.diagramexportmanager:DiagramExportManager",
+            "diagram_export = gaphor.plugins.diagramexport:DiagramExport",
+            "xmi_export = gaphor.plugins.xmiexport:XMIExport",
+            "diagram_layout = gaphor.plugins.diagramlayout:DiagramLayout",
+            "pynsource = gaphor.plugins.pynsource:PyNSource",
+            "alignment = gaphor.plugins.alignment:Alignment",
         ],
     },
     cmdclass={
